Move bootstrap stuff to separate mod
